Re-arrange onboarding-backup-email pages
