+++ assets/i18n/en.json
@@ -251,7 +251,7 @@
     "main.startsIn": "Starts in {{DATE}}",
     "main.suffix": "Suffix",
     "main.tapAboveOnCheckTheCensusAndTryAgain": "Tap above on 'Check the census' and try again",
-    "main.tapSendEmailToSendYourselfABackupLinkForYourRecordsThenClickTheLinkToVerifyThatTheBackupWorks": "Tap \"Send email\" to send yourself a backup link for your records. Then click the link to verify that the backup works.",
+    "main.tapSendEmailToSendYourselfABackupLinkForYourRecordsSoYouCanRecoverTheAccountAtAnyTime": "Tap \"Send email\" to send yourself a backup link for your records so you can recover the account at any time.",
     "main.tapToRetry": "Tap to retry",
     "main.theAddPageAllowsYouToAddLinksCodesAndQRCodesInOrderTo": "The add page allows you to add links, codes and QR codes in order to",
     "main.theCameraIsNotAvailable": "The camera is not available",
@@ -261,6 +261,7 @@
     "main.theMnemonicWordsYouEnteredAreNotValid": "The mnemonic words you entered are not valid",
     "main.theNameContainsInvalidSymbols": "The name contains invalid symbols",
     "main.theNameIsTooShort": "The name is too short",
+    "main.thenClickTheLinkToVerifyThatTheBackupWorks": "Then click the link to verify that the backup works.",
     "main.theNumberOfWordsYouEnteredIsNotValid": "The number of words you entered is not valid",
     "main.thePageCannotBeLoaded": "The page cannot be loaded",
     "main.thePatternNeedsToHaveAtLeastNumDots": "The pattern needs to have at least {{NUM}} dots",
